requirements.txt
langchain
openai
streamlit
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn

